+++ main.h
@@ -4,6 +4,6 @@
 int _putchar(char c);
 int _printf(const char *format, ...);
 int print_number(int n);
-int int_to_bin(unsigned long int nu);
+int int_to_bin(unsigned int nu);
 
 #endif
